+++ plugins/index.js
@@ -3,6 +3,7 @@ class PluginManager {
     this.jira = new (require('./jira'))();
     this.requiredFile = new (require('./required-file'))();
     this.reviewers = new (require('./reviewers'))();
+    this.wip = new (require('./wip'))();
   }
 }